refactor: app modules, main.ts
