--- server/src/apps/immich-admin.main.ts
@@ -1,8 +1,0 @@
-import { CommandFactory } from 'nest-commander';
-import { ImmichAdminModule } from 'src/apps/immich-admin.module';
-import { LogLevel } from 'src/entities/system-config.entity';
-
-export async function bootstrapImmichAdmin() {
-  process.env.LOG_LEVEL = LogLevel.WARN;
-  await CommandFactory.run(ImmichAdminModule);
-}
