+++ server/test/utils.ts
@@ -5,16 +5,14 @@ import fs from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 import { EventEmitter } from 'node:stream';
-import { Server } from 'node:tls';
-import { ApiModule } from 'src/apps/api.module';
-import { ApiService } from 'src/apps/api.service';
-import { AppModule, AppTestModule } from 'src/apps/app.module';
-import { MicroservicesService } from 'src/apps/microservices.service';
+import { AppTestModule } from 'src/app.module';
 import { dataSource } from 'src/database.config';
 import { IJobRepository, JobItem, JobItemHandler, QueueName } from 'src/interfaces/job.interface';
 import { IMediaRepository } from 'src/interfaces/media.interface';
 import { StorageEventType } from 'src/interfaces/storage.interface';
 import { MediaRepository } from 'src/repositories/media.repository';
+import { ApiService } from 'src/services/api.service';
+import { MicroservicesService } from 'src/services/microservices.service';
 import { EntityTarget, ObjectLiteral } from 'typeorm';
 
 export const IMMICH_TEST_ASSET_PATH = process.env.IMMICH_TEST_ASSET_PATH as string;
@@ -104,12 +102,7 @@ let app: INestApplication;
 
 export const testApp = {
   create: async (): Promise<INestApplication> => {
-    const moduleFixture = await Test.createTestingModule({
-      imports: [ApiModule],
-      providers: [ApiService, MicroservicesService],
-    })
-      .overrideModule(AppModule)
-      .useModule(AppTestModule)
+    const moduleFixture = await Test.createTestingModule({ imports: [AppTestModule] })
       .overrideProvider(IJobRepository)
       .useClass(JobMock)
       .overrideProvider(IMediaRepository)
@@ -117,14 +110,10 @@ export const testApp = {
       .compile();
 
     app = await moduleFixture.createNestApplication().init();
-    await app.listen(0);
+    await app.get(ApiService).init();
     await db.reset();
     await app.get(ApiService).init();
     await app.get(MicroservicesService).init();
-
-    const port = app.getHttpServer().address().port;
-    const protocol = app instanceof Server ? 'https' : 'http';
-    process.env.IMMICH_INSTANCE_URL = protocol + '://127.0.0.1:' + port;
 
     return app;
   },
